requirement.txt
Django==4.2.13
djangorestframework==3.15.2
django-filter==24.3
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
python-dotenv==1.0.1
python-telegram-bot==20.6
celery==5.4.0
redis==5.0.7
